automatically save meeting data when user stops recording
