+++ react/features/toolbox/components/web/DownloadDataButton.tsx
@@ -4,43 +4,19 @@ import { connect } from 'react-redux';
 import { IReduxState, IStore } from '../../../app/types';
 import { translate } from '../../../base/i18n/functions';
 import { IconDownload } from '../../../base/icons/svg';
-import { getLocalParticipant, getRemoteParticipants, isLocalParticipantModerator } from '../../../base/participants/functions';
-import { IParticipant } from '../../../base/participants/types';
+import { isLocalParticipantModerator } from '../../../base/participants/functions';
 import AbstractButton from '../../../base/toolbox/components/AbstractButton';
-import { IMessage } from '../../../chat/types';
-import { IPoll } from '../../../polls/types';
+import { downloadMeetingData } from '../../functions.web';
 
 /**
  * The type of the React {@code Component} props of {@link DownloadDataButton}.
  */
 interface IProps extends WithTranslation {
-    _allParticipants: IParticipant[];
-    _localParticipant?: IParticipant;
-    _messages: IMessage[];
-    _polls: { [pollId: string]: IPoll; };
     _room?: string;
+    _state: IReduxState;
     _visible: boolean;
     dispatch: IStore['dispatch'];
 }
-
-/**
- * A helper function to trigger the download of a text file.
- *
- * @param {string} content - The content to be saved in the file.
- * @param {string} filename - The name for the downloaded file.
- * @returns {void}
- */
-const downloadTextFile = (content: string, filename: string) => {
-    const element = document.createElement('a');
-
-    const file = new Blob([ content ]);
-
-    element.href = URL.createObjectURL(file);
-    element.download = filename;
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-};
 
 class DownloadDataButton extends AbstractButton<IProps> {
     override accessibilityLabel = 'toolbar.accessibilityLabel.downloadDataTooltip';
@@ -49,64 +25,12 @@ class DownloadDataButton extends AbstractButton<IProps> {
     override tooltip = 'toolbar.accessibilityLabel.downloadDataTooltip';
 
     /**
-     * Handles clicking the button, and gathers and downloads the data.
+     * Handles clicking the button, and downloads the data.
      *
      * @returns {void}
      */
     override _handleClick() {
-        const { _room, _localParticipant, _allParticipants, _messages, _polls } = this.props;
-
-        const participantNameMap = new Map(_allParticipants.map(p => [ p.id, p.name ]));
-
-        let dataString = `Meeting Data for: ${_room}\n`;
-
-        dataString += `Date: ${new Date().toLocaleString()}\n\n`;
-
-        // 1. Format Attendance List
-        dataString += '--- Attendance ---\n';
-        _allParticipants.forEach(p => {
-            const isLocal = p.id === _localParticipant?.id;
-
-            dataString += `${p.name}${isLocal ? ' (Me)' : ''}\n`;
-        });
-        dataString += '\n';
-
-        // 2. Format Chat History
-        dataString += '--- Chat History ---\n';
-        _messages.forEach(msg => {
-            const displayName = participantNameMap.get((msg as any).participantId) || 'Unknown User';
-            const timestamp = new Date(msg.timestamp).toLocaleTimeString();
-
-            dataString += `[${timestamp}] ${displayName}: ${msg.message}\n`;
-        });
-        dataString += '\n';
-
-        // 3. Format Poll Results
-        dataString += '--- Polls ---\n';
-        if (Object.keys(_polls).length > 0) {
-            Object.values(_polls).forEach((poll, index) => {
-                dataString += `Poll ${index + 1}: ${poll.question}\n`;
-                (poll.answers || []).forEach(answer => {
-                    const voteCount = answer.voters.length;
-                    const voterNames = answer.voters
-                        .map(voterId => participantNameMap.get(voterId) || 'Unknown User')
-                        .join(', ');
-
-                    dataString += `  - ${answer.name} (${voteCount} votes): [${voterNames}]\n`;
-                });
-                dataString += '\n';
-            });
-        } else {
-            dataString += 'No polls were conducted.\n\n';
-        }
-
-        const now = new Date();
-        const year = now.getFullYear();
-        const month = String(now.getMonth() + 1).padStart(2, '0');
-        const day = String(now.getDate()).padStart(2, '0');
-        const dateString = `${year}-${month}-${day}`;
-
-        downloadTextFile(dataString, `${_room}_${dateString}.txt`);
+        downloadMeetingData(this.props._state);
     }
 
     /**
@@ -126,25 +50,10 @@ class DownloadDataButton extends AbstractButton<IProps> {
  * @returns {Object} The mapped props.
  */
 function _mapStateToProps(state: IReduxState) {
-    const localParticipant = getLocalParticipant(state);
-
-    const remoteParticipants = getRemoteParticipants(state);
-
-    const allParticipants: IParticipant[] = [];
-
-    if (localParticipant) {
-        allParticipants.push(localParticipant);
-    }
-    allParticipants.push(...Array.from(remoteParticipants.values()));
-
-
     return {
         _visible: isLocalParticipantModerator(state),
         _room: state['features/base/conference']?.room,
-        _localParticipant: localParticipant,
-        _allParticipants: allParticipants,
-        _messages: state['features/chat'].messages,
-        _polls: state['features/polls']?.polls ?? {}
+        _state: state
     };
 }
